+++ tests/test_NIRSpec.py
@@ -13,9 +13,10 @@ from eureka.S2_calibrations import s2_calibrate as s2
 from eureka.S3_data_reduction import s3_reduce as s3
 from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
+from tests.s4_reference import write_s4_reference_metadata
 
 
-def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
+def test_NIRSpec(capsys, keep_s2_output, keep_s3_output, keep_s4_output):
     with capsys.disabled():
         # is able to display any message without failing a test
         # useful to leave messages for future users who run the tests
@@ -51,6 +52,8 @@ def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
 
     s4_spec, s4_lc, s4_meta = s4.genlc(meta.eventlabel, ecf_path=ecf_path,
                                        s3_meta=s3_meta)
+    if keep_s4_output:
+        write_s4_reference_metadata(s4_meta)
     s5_meta = s5.fitlc(meta.eventlabel, ecf_path=ecf_path, s4_meta=s4_meta)
 
     # run assertions for S2
@@ -99,5 +102,6 @@ def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
                   f"Stage2{os.sep}S2_*")
     if not keep_s3_output:
         os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage3")
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage4")
+    if not keep_s4_output:
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage5")
